Huawei: FileUpload refactoring, fixes Watch 3/4
* replaced fileType with fileId which recived on incoming data
in 28 03 FileHashSend.Response
* width/height in WatchfaceDeviceParams could be int or short
(int is present on Watch 3, short is all other tested devices)
* use in 27 05  WatchfaceConfirm.Request data recieved in previous
WatchfaceConfirm.Response

+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/devices/huawei/HuaweiCoordinator.java
@@ -563,11 +563,11 @@ public class HuaweiCoordinator {
     }
 
     public short getWidth() {
-        return watchfaceDeviceParams.width;
+        return (short)watchfaceDeviceParams.width;
     }
 
     public short getHeight() {
-        return watchfaceDeviceParams.height;
+        return (short)watchfaceDeviceParams.height;
     }
 
     public void setWatchfaceDeviceParams(Watchface.WatchfaceDeviceParams watchfaceDeviceParams) {
